requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9

# PDF Processing
PyMuPDF>=1.24.0
tiktoken>=0.7.0

# Embeddings & Reranking (local)
sentence-transformers>=3.0.0

# LLM — Gemini (optional, only if using gemini provider)
google-genai>=1.0.0

# Vector Store
chromadb>=0.5.0

# Retrieval
rank-bm25>=0.2.2

# Frontend
streamlit>=1.38.0
requests>=2.32.0

# Dev & Utilities
python-dotenv>=1.0.0
pytest>=8.0
ruff>=0.6.0
